* use size_t instead unsigned int for Glib::ustring::iterator  * fix wrong var for mono convolver controller  * add check for boost_system in wscript

+++ trunk/src/gx_head/engine/ladspaplugin.cpp
@@ -291,11 +291,11 @@ void LadspaDsp::stereo_process(int count, float *input1, float *input2, float *o
 }
 
 static Glib::ustring TrimLabel(const char *label, int cnt_in_row) {
-    const unsigned int minlen = 60  / cnt_in_row - 1;
-    const unsigned int maxlen = minlen + 10;
-    const unsigned int cutlen = (maxlen + minlen) / 2;
+    const size_t minlen = 60  / cnt_in_row - 1;
+    const size_t maxlen = minlen + 10;
+    const size_t cutlen = (maxlen + minlen) / 2;
     Glib::ustring pn(label);
-    unsigned int rem = pn.find_first_of("([");
+    size_t rem = pn.find_first_of("([");
     if(rem != Glib::ustring::npos) {
 	pn.erase(rem);
     }
@@ -303,8 +303,8 @@ static Glib::ustring TrimLabel(const char *label, int cnt_in_row) {
 	pn.erase(rem);
     }
     rem = 0;
-    unsigned int rem1 = 0;
-    unsigned int lastpos = 0;
+    size_t rem1 = 0;
+    size_t lastpos = 0;
     while (true) {
 	rem1 = pn.find_first_of(" ", rem1);
 	if (rem1 == Glib::ustring::npos) {
